requirements.txt
Django==1.4.1
dj-database-url==0.2.1
psycopg2==2.4.5
wsgiref==0.1.2
gunicorn==0.14.6
South==0.7.5
django-secure==0.1.2
pytz==2012d
django-orm-extensions==3.0b3
django-celery==3.0.6
django-braces==0.1.3
django-debug-toolbar==0.9.4
django-floppyforms==1.0
mock==1.0b1
django-registration==0.8
Sphinx==1.1.3
